# WebAPI/backend/app/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/ai/generate_description":{"post":{"summary":"Generate Description","operationId":"generate_description_ai_generate_description_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InputData"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/ai/generate_tags":{"post":{"summary":"Generate Tags Endpoint","operationId":"generate_tags_endpoint_ai_generate_tags_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InputData"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/ai/categorize_ewaste_base64":{"post":{"summary":"Categorize E Waste Base64","operationId":"categorize_e_waste_base64_ai_categorize_ewaste_base64_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ImageData"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ImageData":{"properties":{"image_base64":{"type":"string","title":"Image Base64"}},"type":"object","required":["image_base64"],"title":"ImageData"},"InputData":{"properties":{"user_input":{"type":"string","title":"User Input"}},"type":"object","required":["user_input"],"title":"InputData"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"token"}}}}}}